Remove legacy WorkoutPlanner and associated tests
Removed `WorkoutPlanner.kt` and its test file as the legacy placeholder logic is no longer needed.
Extracted `WorkoutStep` data class to its own file `WorkoutStep.kt` to preserve the data model used by `WorkoutViewModel`.
Updated `WorkoutViewModel` to remove the fallback call to the legacy planner, now defaulting to an empty list if the advanced planner is not available.
Cleaned up unused imports in `ExerciseSelectionScreen`.
Updated `WorkoutViewModelTest` to mock dependencies correctly for testing active states, removing reliance on the legacy fallback.
